Improve docs

+++ PrepatcherPlugin/EventSubscriptionPriorityAttribute.cs
@@ -15,14 +15,14 @@ public class EventSubscriptionPriorityAttribute : Attribute
     /// <summary>
     /// The priority of the subscriber.
     /// </summary>
-    public int Priority { get; set; }
+    public float Priority { get; set; }
 
     /// <summary>
     /// Create an instance of this attribute with the specified priority.
     /// </summary>
-    public EventSubscriptionPriorityAttribute(int priority) => Priority = priority;
+    public EventSubscriptionPriorityAttribute(float priority) => Priority = priority;
 
-    internal static int GetPriority<T>(T? func) where T : Delegate
+    internal static float GetPriority<T>(T? func) where T : Delegate
     {
         return func?.Method?.GetCustomAttribute<EventSubscriptionPriorityAttribute>()?.Priority ?? 0;
     }
